Add lustre support to MPIFileUtils
`dcp` has a `--grouplock` option and that requires the lustre client.
Lustre is now being built in the builder stage of the image and is used
to enable lustre build support in MPIFileUtils and also runtime support.

This image's OS is debian bookworm, which has deprecated `dpatch` - a
required build dependency for older versions of lustre. The latest cray
release of lustre 2.15 is being used, where `dpatch` is no longer
required.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>